Fixed all but one comment
