+++ app/[lang]/layout.tsx
@@ -2,12 +2,17 @@ import type { Metadata } from 'next';
 import { ReactNode } from 'react';
 import ThemeRegistry from '@/app/components/ThemeRegistry';
 import { Box } from '@mui/material';
+import { Inter } from 'next/font/google';
 
 export const metadata: Metadata = {
   title: 'Create Next App',
   description: 'Generated by create next app',
 };
 
+const inter = Inter({
+  subsets: ['latin'],
+  display: 'swap',
+});
 export default function RootLayout({
   children,
   params: { lang },
@@ -16,7 +21,7 @@ export default function RootLayout({
   params: { lang: string };
 }) {
   return (
-    <html lang={lang}>
+    <html lang={lang} className={inter.className}>
       <head>
         <link rel="manifest" href="/manifest.json" />
         <meta name="theme-color" content="#90cdf4" />
