+++ app/[lang]/not-found.tsx
@@ -1,0 +1,13 @@
+import Link from 'next/link';
+
+const NotFound = () => {
+  return (
+    <div>
+      <h2>Not Found</h2>
+      <p>Could not find requested resource</p>
+      <Link href="/">Return Home</Link>
+    </div>
+  );
+};
+
+export default NotFound;
