+++ app/components/onboarding/AddNewSiteDialog.tsx
@@ -11,6 +11,7 @@ import {
   Typography,
 } from '@mui/material';
 import { useState } from 'react';
+import { Dict } from '@/app/[lang]/dictionaries';
 
 interface OnboardingAddNewSiteDialogProps {
   dict: Dict;
